revert test script names
